undo delete add

+++ lib/blocs/note_bloc/note_bloc.dart
@@ -17,7 +17,8 @@ part 'note_state.dart';
 class NoteBloc extends Bloc<NoteEvent, NoteState> {
   NoteRepository repository = NoteService();
   List<Note> _notes = [];
-
+  Note? _lastRemovedNote;
+  bool isMore=true;
   NoteBloc() : super(NoteInitial()) {
     on<InitRepository>((event, emit) async {
       emit(NoteLoading());
@@ -39,11 +40,11 @@ class NoteBloc extends Bloc<NoteEvent, NoteState> {
         onCreate: (db, version) async {
           await db.execute('''
             CREATE TABLE notes(
-              id TEXT PRIMARY KEY,
+              id TEXT,
               title TEXT,
               content TEXT,
-              userId TEXT,
-              createdAt TEXT
+              user_id TEXT,
+              created_at TEXT
             )
           ''');
         },
@@ -56,10 +57,22 @@ class NoteBloc extends Bloc<NoteEvent, NoteState> {
       add(LoadNotes());
     });
     on<LoadNotes>((event, emit) async {
-      emit(NoteLoading());
+      if(!event.isMore) {
+       _notes.clear();
+        emit(NoteLoading());
+      }
+      emit(NoteLoaded(_notes,isMore: isMore,isLoading: true));
+
       try {
-        _notes = await repository.getNotes(search: event.search);
-        emit(NoteLoaded(_notes));
+        List<Note> notes = await repository.getNotes(search: event.search, nextToken: event.nextToken);
+        if(notes.isEmpty){
+          isMore=false;
+        }
+        for (var data in notes) {
+          _notes.remove(data);
+          _notes.add(data);
+        }
+        emit(NoteLoaded(_notes,isMore: isMore,));
       } catch (e) {
         emit(NoteError(e.toString()));
       }
@@ -69,7 +82,7 @@ class NoteBloc extends Bloc<NoteEvent, NoteState> {
       try {
         final newNote = await repository.createNote(event.note);
         _notes.insert(0, newNote);
-        emit(NoteLoaded(List.from(_notes)));
+        emit(NoteLoaded(List.from(_notes),isMore: isMore));
       } catch (e) {
         emit(NoteError(e.toString()));
       }
@@ -79,7 +92,7 @@ class NoteBloc extends Bloc<NoteEvent, NoteState> {
       try {
         final updatedNote = await repository.updateNote(event.note);
         _notes = _notes.map((n) => n.id == event.note.id ? updatedNote : n).toList();
-        emit(NoteLoaded(List.from(_notes)));
+        emit(NoteLoaded(List.from(_notes),isMore: isMore));
       } catch (e) {
         emit(NoteError(e.toString()));
       }
@@ -87,11 +100,29 @@ class NoteBloc extends Bloc<NoteEvent, NoteState> {
 
     on<RemoveNote>((event, emit) async {
       try {
-        await repository.deleteNote(event.id);
+        final removedNote = _notes.firstWhere((n) => n.id == event.id);
+
         _notes.removeWhere((n) => n.id == event.id);
-        emit(NoteLoaded(List.from(_notes)));
+        emit(NoteLoaded(List.from(_notes),isMore: isMore));
+
+        Future.delayed(const Duration(seconds: 30), () async {
+          final stillRemoved = !_notes.any((n) => n.id == removedNote.id);
+          if (stillRemoved) {
+            await repository.deleteNote(event.id);
+          }
+        });
+
+        _lastRemovedNote = removedNote;
       } catch (e) {
         emit(NoteError(e.toString()));
+      }
+    });
+
+    on<UndoRemoveNote>((event, emit) {
+      if (_lastRemovedNote != null) {
+        _notes.add(_lastRemovedNote!);
+        emit(NoteLoaded(List.from(_notes),isMore: isMore));
+        _lastRemovedNote = null;
       }
     });
   }
